requirements.txt
azure-identity~=1.15.0
msgraph-sdk~=1.2.0